int64

+++ database.go
@@ -2,7 +2,7 @@ package core
 
 type Database struct {
 	Connections map[string]DatabaseSettings `json:"connections,omitempty"`
-	PingTimeout uint32                      `json:"pingTimeout,omitempty"`
+	PingTimeout int64                       `json:"pingTimeout,omitempty"`
 }
 
 type DatabaseSettings struct {
