+++ redis.go
@@ -2,6 +2,6 @@ package core
 
 type Redis struct {
 	Host   string `json:"host"`
-	Port   uint32 `json:"port"`
+	Port   int64 `json:"port"`
 	Prefix string `json:"prefix"`
 }
